commit 2

+++ main.c
@@ -3,8 +3,8 @@
 #include <net/ethernet.h>
 #include <netinet/ip.h>
 #include <netinet/ip_icmp.h>
-#include <arpa/inet.h> // Thêm để khai báo inet_ntoa
-#include <libnetfilter_queue/linux_nfnetlink_queue.h> // Thêm để khai báo NF_ACCEPT
+#include <arpa/inet.h> // Để khai báo inet_ntoa
+#include <libnetfilter_queue/linux_nfnetlink_queue.h> // Để khai báo NF_ACCEPT
 
 static int cb(struct nfq_q_handle *qh, struct nfgenmsg *nfmsg, struct nfq_data *nfa, void *data) {
     struct nfqnl_msg_packet_hdr *ph = nfq_get_msg_packet_hdr(nfa);
@@ -15,10 +15,9 @@ static int cb(struct nfq_q_handle *qh, struct nfgenmsg *nfmsg, struct nfq_data *
         unsigned char *payload;
         int payload_len = nfq_get_payload(nfa, &payload);
 
-        // Ưu tiên kiểm tra header IP (trường hợp không có Ethernet)
         if (payload_len >= sizeof(struct iphdr)) {
             struct iphdr *ip = (struct iphdr *)payload;
-            int ip_header_len = ip->ihl * 4; // Độ dài header IP thực tế
+            int ip_header_len = ip->ihl * 4;
             if (ip->version == 4 && ip->ihl >= 5 && ip_header_len <= payload_len) {
                 uint16_t ip_total_len = ntohs(ip->tot_len);
                 printf("Bắt đầu từ header IP tại offset 0, Nguồn: %s, Độ dài header IP: %d byte, Tổng độ dài IP: %u byte\n",
@@ -27,18 +26,16 @@ static int cb(struct nfq_q_handle *qh, struct nfgenmsg *nfmsg, struct nfq_data *
                 if (payload_len >= ip_total_len) {
                     printf("Dữ liệu đầy đủ: %d byte\n", payload_len);
                 } else {
-                    printf("Dữ liệu không đầy đủ, payload_len: %d, ip_total_len: %u\n",
-                           payload_len, ip_total_len);
+                    printf("Dữ liệu không đầy đủ, payload_len: %d, ip_total_len: %lu\n",
+                           payload_len, (unsigned long)ip_total_len);
                 }
 
-                // Kiểm tra ICMP (Ping)
                 if (ip->protocol == IPPROTO_ICMP && payload_len >= ip_header_len + sizeof(struct icmphdr)) {
                     struct icmphdr *icmp = (struct icmphdr *)(payload + ip_header_len);
                     printf("Gói tin ICMP, Type: %d (0=Reply, 8=Request), Code: %d\n",
                            icmp->type, icmp->code);
                     if (icmp->type == 8 || icmp->type == 0) {
                         printf("Đây là bản tin Ping (Echo %s)\n", icmp->type == 8 ? "Request" : "Reply");
-                        // In một phần dữ liệu ICMP nếu có (ví dụ 8 byte đầu)
                         printf("Dữ liệu ICMP đầu tiên: ");
                         for (int i = 0; i < 8 && i < (payload_len - ip_header_len - sizeof(struct icmphdr)); i++) {
                             printf("%02x ", *(payload + ip_header_len + sizeof(struct icmphdr) + i));
@@ -47,7 +44,6 @@ static int cb(struct nfq_q_handle *qh, struct nfgenmsg *nfmsg, struct nfq_data *
                     }
                 }
 
-                // Kiểm tra header Ethernet nếu đủ dài
                 if (payload_len >= sizeof(struct ethhdr) + ip_total_len) {
                     struct ethhdr *eth = (struct ethhdr *)payload;
                     if (ntohs(eth->h_proto) == ETH_P_IP) {
@@ -72,11 +68,10 @@ static int cb(struct nfq_q_handle *qh, struct nfgenmsg *nfmsg, struct nfq_data *
                     if (payload_len >= sizeof(struct ethhdr) + ip_total_len) {
                         printf("Dữ liệu đầy đủ: %d byte\n", payload_len);
                     } else {
-                        printf("Dữ liệu không đầy đủ, payload_len: %d, ip_total_len: %u\n",
-                               payload_len, ip_total_len + sizeof(struct ethhdr));
+                        printf("Dữ liệu không đầy đủ, payload_len: %d, ip_total_len: %lu\n",
+                               payload_len, (unsigned long)(ip_total_len + sizeof(struct ethhdr)));
                     }
 
-                    // Kiểm tra ICMP (Ping)
                     if (ip->protocol == IPPROTO_ICMP && payload_len >= sizeof(struct ethhdr) + ip_header_len + sizeof(struct icmphdr)) {
                         struct icmphdr *icmp = (struct icmphdr *)(payload + sizeof(struct ethhdr) + ip_header_len);
                         printf("Gói tin ICMP, Type: %d (0=Reply, 8=Request), Code: %d\n",
